quest fix update
- fixed text position on success event
