+++ Clicker/Clicker.csproj
@@ -222,8 +222,10 @@
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
-    <Compile Include="Models\AchievementModel.cs" />
-    <Compile Include="Models\UpgradeModel.cs" />
+    <Compile Include="Models\Achievement.cs" />
+    <Compile Include="Models\GameSettings.cs" />
+    <Compile Include="Models\Player.cs" />
+    <Compile Include="Models\Upgrade.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
