+++ Clicker/Content/Site.css
@@ -85,3 +85,14 @@ div {
     padding-top: 3px;
     text-align: center;
 }
+
+ .progress {
+     position: relative;
+ }
+
+ .progress span {
+    position: absolute;
+    display: block;
+    width: 100%;
+    color: white;
+ }
