--- Clicker/Models/AchievementModel.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace Clicker.Models
-{
-    public class Achievement
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public int Level { get; set; }
-        public int Value { get; set; }
-        public int Cost { get; set; }
-        public string ImgUrl { get; set; }
-        public string Description { get; set; }
-
-    }
-}
